chore: use HyperlaneDb trait in validator submitter
